vtx: split file for sa and tramp
