+++ src/io/vtx.c
@@ -19,16 +19,6 @@
 static int fpv_init = 0;
 #endif
 
-#define VTX_SETTINGS_MAGIC 0xdeed
-#define VTX_APPLY_TRIES 10
-
-typedef enum {
-  VTX_DETECT_WAIT,
-  VTX_DETECT_ERROR,
-  VTX_DETECT_SUCCESS,
-  VTX_DETECT_UPDATE
-} vtx_detect_status_t;
-
 vtx_settings_t vtx_settings;
 uint8_t vtx_connect_tries = 0;
 
@@ -36,6 +26,22 @@ static vtx_settings_t vtx_actual;
 
 static uint32_t vtx_delay_start = 0;
 static uint32_t vtx_delay_ms = 1000;
+
+extern uint8_t smart_audio_detected;
+extern smart_audio_settings_t smart_audio_settings;
+
+vtx_detect_status_t vtx_smart_audio_update(vtx_settings_t *actual);
+void smart_audio_set_frequency(vtx_band_t band, vtx_channel_t channel);
+void smart_audio_set_power_level(vtx_power_level_t power);
+void smart_audio_set_pit_mode(vtx_pit_mode_t pit_mode);
+
+extern uint8_t tramp_detected;
+extern tramp_settings_t tramp_settings;
+
+vtx_detect_status_t vtx_tramp_update(vtx_settings_t *actual);
+void tramp_set_frequency(vtx_band_t band, vtx_channel_t channel);
+void tramp_set_power_level(vtx_power_level_t power);
+void tramp_set_pit_mode(vtx_pit_mode_t pit_mode);
 
 const uint16_t frequency_table[VTX_BAND_MAX][VTX_CHANNEL_MAX] = {
     {5865, 5845, 5825, 5805, 5785, 5765, 5745, 5725}, // VTX_BAND_A
@@ -59,238 +65,6 @@ int8_t vtx_find_frequency_index(uint16_t frequency) {
   }
   return -1;
 }
-
-#ifdef ENABLE_SMART_AUDIO
-
-#define SMART_AUDIO_DETECT_TRIES 30
-
-extern smart_audio_settings_t smart_audio_settings;
-uint8_t smart_audio_detected = 0;
-static bool smart_audio_needs_update = false;
-
-vtx_detect_status_t vtx_smart_audio_update(vtx_settings_t *actual) {
-  if (smart_audio_settings.version == 0 && vtx_connect_tries > SMART_AUDIO_DETECT_TRIES) {
-    return VTX_DETECT_ERROR;
-  }
-
-  const vtx_update_result_t result = serial_smart_audio_update();
-
-  if ((result == VTX_IDLE || result == VTX_ERROR) && smart_audio_settings.version == 0 && vtx_connect_tries <= SMART_AUDIO_DETECT_TRIES) {
-    // no smart audio detected, try again
-    serial_smart_audio_send_payload(SA_CMD_GET_SETTINGS, NULL, 0);
-    vtx_connect_tries++;
-    return VTX_DETECT_WAIT;
-  }
-
-  if (result == VTX_SUCCESS) {
-    int8_t channel_index = -1;
-    if (smart_audio_settings.mode & SA_MODE_FREQUENCY) {
-      channel_index = vtx_find_frequency_index(smart_audio_settings.frequency);
-    } else {
-      channel_index = smart_audio_settings.channel;
-    }
-
-    if (channel_index >= 0) {
-      actual->band = channel_index / VTX_CHANNEL_MAX;
-      actual->channel = channel_index % VTX_CHANNEL_MAX;
-    }
-
-    if (smart_audio_settings.version == 2) {
-      actual->power_level = min(smart_audio_settings.power, VTX_POWER_LEVEL_4);
-    } else {
-      actual->power_level = smart_audio_dac_power_level_index(smart_audio_settings.power);
-    }
-
-    if (smart_audio_settings.version >= 3) {
-      actual->pit_mode = smart_audio_settings.mode & SA_MODE_PIT ? 1 : 0;
-    } else {
-      actual->pit_mode = VTX_PIT_MODE_NO_SUPPORT;
-    }
-
-    if (smart_audio_settings.version != 0 && smart_audio_detected == 0) {
-      quic_debugf("smart audio version: %d", smart_audio_settings.version);
-      smart_audio_detected = 1;
-
-      if (vtx_settings.magic != VTX_SETTINGS_MAGIC) {
-        vtx_set(actual);
-      }
-
-      vtx_settings.detected = VTX_PROTOCOL_SMART_AUDIO;
-      vtx_connect_tries = 0;
-    }
-
-    if (smart_audio_needs_update) {
-      serial_smart_audio_send_payload(SA_CMD_GET_SETTINGS, NULL, 0);
-      smart_audio_needs_update = false;
-      return VTX_DETECT_WAIT;
-    }
-
-    return VTX_DETECT_SUCCESS;
-  }
-
-  if ((result == VTX_IDLE || result == VTX_ERROR) && smart_audio_detected) {
-    // we are detected and vtx is in idle, we can update stuff
-    return VTX_DETECT_UPDATE;
-  }
-
-  // wait otherwise
-  return VTX_DETECT_WAIT;
-}
-
-static void smart_audio_set_frequency(vtx_band_t band, vtx_channel_t channel) {
-#ifdef SA_USE_SET_FREQUENCY
-  if (smart_audio_settings.mode & SA_MODE_FREQUENCY) {
-    const uint16_t frequency = frequency_table[band][channel];
-    const uint8_t payload[2] = {
-        (frequency >> 8) & 0xFF,
-        frequency & 0xFF,
-    };
-    serial_smart_audio_send_payload(SA_CMD_SET_FREQUENCY, payload, 2);
-  } else
-#endif
-  {
-    smart_audio_settings.mode &= ~SA_MODE_FREQUENCY;
-
-    const uint8_t index = band * VTX_CHANNEL_MAX + channel;
-    const uint8_t payload[1] = {index};
-    serial_smart_audio_send_payload(SA_CMD_SET_CHANNEL, payload, 1);
-  }
-}
-
-static void smart_audio_set_power_level(vtx_power_level_t power) {
-  uint8_t level = power;
-  if (smart_audio_settings.version != 2) {
-    level = smart_audio_settings.dac_power_levels[power];
-  }
-  if (smart_audio_settings.version == 3) {
-    level |= 0x80;
-  }
-  serial_smart_audio_send_payload(SA_CMD_SET_POWER, &level, 1);
-}
-
-static void smart_audio_set_pit_mode(vtx_pit_mode_t pit_mode) {
-  if (smart_audio_settings.version >= 3) {
-    uint8_t mode = 0x0;
-
-    if (pit_mode == VTX_PIT_MODE_OFF) {
-      mode |= 0x04;
-    } else if (pit_mode == VTX_PIT_MODE_ON) {
-      // out-range was dropped for VTXes with SA >= v2.1
-      mode |= 0x01;
-    }
-
-    serial_smart_audio_send_payload(SA_CMD_SET_MODE, &mode, 1);
-    smart_audio_needs_update = true;
-  } else {
-    vtx_settings.pit_mode = VTX_PIT_MODE_NO_SUPPORT;
-  }
-}
-
-#endif
-
-#ifdef ENABLE_TRAMP
-
-#define TRAMP_DETECT_TRIES 5
-
-extern tramp_settings_t tramp_settings;
-uint8_t tramp_detected = 0;
-
-const uint16_t tramp_power_level[4] = {
-    25,
-    100,
-    200,
-    400,
-};
-
-vtx_power_level_t tramp_power_level_index(uint16_t power) {
-  for (uint8_t level = 0; level < VTX_POWER_LEVEL_MAX; level++) {
-    if (power >= tramp_power_level[level] && power <= tramp_power_level[level]) {
-      return level;
-    }
-  }
-  return VTX_POWER_LEVEL_1;
-}
-
-vtx_detect_status_t vtx_tramp_update(vtx_settings_t *actual) {
-  if (tramp_settings.freq_min == 0 && vtx_connect_tries > TRAMP_DETECT_TRIES) {
-    return VTX_DETECT_ERROR;
-  }
-
-  vtx_update_result_t result = serial_tramp_update();
-
-  if ((result == VTX_IDLE || result == VTX_ERROR) && tramp_settings.freq_min == 0 && vtx_connect_tries <= TRAMP_DETECT_TRIES) {
-    // no tramp detected, try again
-    serial_tramp_send_payload('r', 0);
-    vtx_connect_tries++;
-    return VTX_DETECT_WAIT;
-  }
-
-  if (result == VTX_SUCCESS) {
-    if (tramp_settings.freq_min == 0) {
-      // tramp reset was successful but returned empty data
-      return VTX_DETECT_WAIT;
-    }
-
-    if (tramp_settings.frequency == 0) {
-      serial_tramp_send_payload('v', 0);
-      return VTX_DETECT_WAIT;
-    }
-
-    int8_t channel_index = vtx_find_frequency_index(tramp_settings.frequency);
-    if (channel_index >= 0) {
-      actual->band = channel_index / VTX_CHANNEL_MAX;
-      actual->channel = channel_index % VTX_CHANNEL_MAX;
-    }
-
-    actual->power_level = tramp_power_level_index(tramp_settings.power);
-    actual->pit_mode = tramp_settings.pit_mode;
-
-    // not all vtxes seem to return a non-zero value. :(
-    // as its unused lets just drop it.
-    // if (tramp_settings.temp == 0) {
-    //   serial_tramp_send_payload('s', 0);
-    //   return VTX_DETECT_WAIT;
-    // }
-
-    if (tramp_settings.freq_min != 0 && tramp_settings.frequency != 0 && tramp_detected == 0) {
-      tramp_detected = 1;
-
-      if (vtx_settings.magic != VTX_SETTINGS_MAGIC) {
-        vtx_set(actual);
-      }
-
-      vtx_settings.detected = VTX_PROTOCOL_TRAMP;
-      vtx_connect_tries = 0;
-    }
-    return VTX_DETECT_SUCCESS;
-  }
-
-  if ((result == VTX_IDLE || result == VTX_ERROR) && tramp_detected) {
-    // we are detected and vtx is in idle, we can update stuff
-    return VTX_DETECT_UPDATE;
-  }
-
-  // wait otherwise
-  return VTX_DETECT_WAIT;
-}
-
-static void tramp_set_frequency(vtx_band_t band, vtx_channel_t channel) {
-  const uint16_t frequency = frequency_table[band][channel];
-  serial_tramp_send_payload('F', frequency);
-  tramp_settings.frequency = 0;
-}
-
-static void tramp_set_power_level(vtx_power_level_t power) {
-  serial_tramp_send_payload('P', tramp_power_level[power]);
-  tramp_settings.frequency = 0;
-}
-
-static void tramp_set_pit_mode(vtx_pit_mode_t pit_mode) {
-  serial_tramp_send_payload('I', pit_mode == VTX_PIT_MODE_ON ? 0 : 1);
-  tramp_settings.frequency = 0;
-}
-
-#endif
 
 static vtx_detect_status_t vtx_update_protocol(vtx_protocol_t proto, vtx_settings_t *actual) {
   switch (proto) {
