+++ src/io/vtx.h
@@ -2,6 +2,16 @@
 
 #include <cbor.h>
 
+#define VTX_SETTINGS_MAGIC 0xdeed
+#define VTX_APPLY_TRIES 10
+
+typedef enum {
+  VTX_DETECT_WAIT,
+  VTX_DETECT_ERROR,
+  VTX_DETECT_SUCCESS,
+  VTX_DETECT_UPDATE
+} vtx_detect_status_t;
+
 typedef enum {
   VTX_BAND_A,
   VTX_BAND_B,
@@ -71,6 +81,7 @@ void vtx_update();
 void vtx_set(vtx_settings_t *vtx);
 
 uint16_t vtx_frequency_from_channel(vtx_band_t band, vtx_channel_t channel);
+int8_t vtx_find_frequency_index(uint16_t frequency);
 
 cbor_result_t cbor_encode_vtx_settings_t(cbor_value_t *enc, const vtx_settings_t *vtx);
 cbor_result_t cbor_decode_vtx_settings_t(cbor_value_t *dec, vtx_settings_t *vtx);
